requirements.txt
flask
flask-sqlalchemy
flask-cors
python-dotenv
google-generativeai